+++ static/primary.json
@@ -2,49 +2,63 @@
     "nodes": [
         {
             "handle": "work",
-            "label": "Work/Variant",
+            "label_en": "Work/Variant",
+            "label_es": "Obra/Variante",
+            "label_fr": "l'Œuvre/la Variante",
             "url": "WorkVariant",
             "x": 100,
             "y": 50
         },
         {
             "handle": "manifestation",
-            "label": "Manifestation",
+            "label_en": "Manifestation",
+            "label_es": "Manifestación",
+            "label_fr": "Manifestation",
             "url": "Manifestation",
             "x": 100,
             "y": 120
         },
         {
             "handle": "item",
-            "label": "Item",
+            "label_en": "Item",
+            "label_es": "Ítem",
+            "label_fr": "Item",
             "url": "Item",
             "x": 100,
             "y": 190
         },
         {
             "handle": "carrier",
-            "label": "Carrier",
+            "label_en": "Carrier",
+            "label_es": "Portador",
+            "label_fr": "Article",
             "url": "Carrier",
             "x": 100,
             "y": 260
         },
         {
             "handle": "event",
-            "label": "Event",
+            "label_en": "Event",
+            "label_es": "Evento",
+            "label_fr": "Événement",
             "url": "Event",
             "x": 250,
             "y": 155
         },
         {
             "handle": "activity",
-            "label": "Activity",
+            "label_en": "Activity",
+            "label_es": "Actividad",
+            "label_fr": "Activité",
             "url": "Activity",
             "x": 350,
             "y": 155
         },
         {
             "handle": "agent",
-            "label": "Agent",
+            "label_en": "Agent",
+            "label_es": "Agente",
+            "label_fr": "Agent",
             "url": "Agent",
             "x": 450,
             "y": 155
@@ -53,48 +67,39 @@
     "links": [
         {
             "source": "work",
-            "target": "manifestation",
-            "label": "Has Manfestation"
+            "target": "manifestation"
         },
         {
             "source": "manifestation",
-            "target": "item",
-            "label": "Has Item"
+            "target": "item"
         },
         {
             "source": "item",
-            "target": "carrier",
-            "label": "Has Carrier"
+            "target": "carrier"
         },
         {
             "source": "work",
-            "target": "event",
-            "label": "Has Event"
+            "target": "event"
         },
         {
             "source": "manifestation",
-            "target": "event",
-            "label": "Has Event"
+            "target": "event"
         },
         {
             "source": "item",
-            "target": "event",
-            "label": "Has Event"
+            "target": "event"
         },
         {
             "source": "carrier",
-            "target": "event",
-            "label": "Has Event"
+            "target": "event"
         },
         {
             "source": "event",
-            "target": "activity",
-            "label": "Has Activity"
+            "target": "activity"
         },
         {
             "source": "activity",
-            "target": "agent",
-            "label": "Has Agent"
+            "target": "agent"
         }
     ]
 }
