+++ templates/index_en.html
@@ -124,8 +124,8 @@
         </sub>
     </div>
     <script src="https://d3js.org/d3.v7.min.js"></script>
-    <script>let graph_terms = JSON.parse('{{ data|tojson() }}')</script>
-    <script src="../static/summary.js"></script>
+    <script>let lang = '{{ lang }}'</script>
+    <script src="../static/primary.js"></script>
 
 </body>
